+++ framework/CHANGELOG.md
@@ -47,6 +47,7 @@ Yii Framework 2 Change Log
 - Chg #15448: Package "ezyang/htmlpurifier" has been made optional and is not installed by default (klimov-paul)
 - Chg #15481: Removed `yii\BaseYii::powered()` method (Kolyunya, samdark)
 - Chg #15811: Fixed issue with additional parameters on `yii\base\View::renderDynamic()` while parameters contains single quote introduced in #12938 (xicond)
+- Enh #12147: `yii\db\BaseActiveRecord::bindModels()` no longer saves the foreign record (dizeee)
 
 2.0.14.2 under development
 ------------------------